+++ src/commands/mod.rs
@@ -1,3 +1,4 @@
 pub mod clear_cmd;
+pub mod config_cmd;
 pub mod gen_cmd;
 pub mod new_cmd;